+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/IsoLanguage.kt
@@ -1,0 +1,8 @@
+package com.mapbox.search.base
+
+import com.mapbox.search.common.IsoLanguage
+import java.util.Locale
+
+fun defaultLocaleLanguage(): IsoLanguage {
+    return IsoLanguage(Locale.getDefault().language)
+}