+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/analytics/events/SearchFeedbackEventsFactoryTest.kt
@@ -4,7 +4,6 @@ import android.graphics.Bitmap
 import com.mapbox.geojson.BoundingBox
 import com.mapbox.geojson.Point
 import com.mapbox.search.BuildConfig
-import com.mapbox.search.common.IsoCountry
 import com.mapbox.search.EtaType
 import com.mapbox.search.QueryType
 import com.mapbox.search.ResponseInfo
@@ -31,6 +30,7 @@ import com.mapbox.search.base.utils.FormattedTimeProvider
 import com.mapbox.search.base.utils.UUIDProvider
 import com.mapbox.search.base.utils.extension.mapToPlatform
 import com.mapbox.search.base.utils.orientation.ScreenOrientation
+import com.mapbox.search.common.IsoCountry
 import com.mapbox.search.common.IsoLanguage
 import com.mapbox.search.common.createTestCoreSearchResponseSuccess
 import com.mapbox.search.internal.bindgen.FeedbackEventCallback