+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [CORE] `OpenHours`, `OpenPeriod`, `WeekDay`, `WeekTimestamp`, `ParkingData` classes have been moved to `com.mapbox.search.common.metadata` package.
 - [CORE] `Language` has been renamed to `IsoLanguage` and moved to `com.mapbox.search.common` package.
 - [CORE] `Country` has been renamed to `IsoCountry` and moved to `com.mapbox.search.common` package.
+- [Autofill] `AddressAutofillOptions.Country` and `AddressAutofillOptions.Language` have been removed and replaced by `IsoCountry` and `IsoLanguage`.
 
 ### Mapbox dependencies
 - Search Native SDK `0.66.0`